Change list of events which may be filtered by conditional
It doesn't make sense to present users with a choice to filer by Sun30 events during
the Discipline Era, or during Short Circuits. By including descriptor information in
the outcome definitions, specifically under which weather an event can happen, or in
which seasons an event occured (specifically for help with Elsewhere, Voicemail and
Fax Machine), the list of outcomes presented to the user in the picker can be filtered
to only allow those which make sense.

A bypass conditional is also included, to prevent season filtering from applying to
seasons which have yet to occur, which would by-necessity not be included in the
descriptors which we currently have.

* Reformats the outcome file to put data on multiple lines -- Prettier said so